requirements.txt
opencv-python==4.5.5.64
numpy==1.21.5
ultralytics==8.3.14
pillow==8.4.0